Bring k3s_1_27 back
This is basically a revert of commit 7e891c65c6cd85848a8c762204be2a0aa727f140
with modifications in all-packages.nix. It isn't strictly revertable
because multiple changes happened for other or all k3s versions after that
commit. We keep the current upstream nixpkgs code for newer versions but
bring back the original import for k3s 1.27 from before that commit
because I'm not sure if k3s 1.27 would work properly with the newer Go
version and we want to keep it running as it is now.

The clusters using the outdated version should be upgraded asap as 1.27
is EOL for some months now.

PL-133043